feat: Update speaker morphing weights to use float arrays across the codebase for improved precision and consistency

+++ app/src/main/cpp/beatriceParameters.h
@@ -20,8 +20,8 @@ struct BeatriceParameters {
   int vqNumNeighbors = 4;
   std::array<double, beatrice::common::kMaxNSpeakers + 1>
       averageTargetPitchBase = {0.0};
-  std::array<double, beatrice::common::kMaxNSpeakers> speakerMorphingWeights = {
-      0.0};
+  std::array<float, beatrice::common::kMaxNSpeakers> speakerMorphingWeights = {
+      0.0f};
 };
 
 #endif  // BEATRICE_PARAMETERS_H
